refactor(kernel): remove entity registry compatibility projection
