+++ tools/gate-allowlists/check-kernel-dead-exports.json
@@ -382,24 +382,6 @@
         "until": "2026-09-30"
       },
       {
-        "value": "entityRegistry",
-        "ref": "task_01KZRMKJT7YKQ2TNXA8WR1W5DN",
-        "reason": "W3 production flip removed the former CLI barrel consumer; the private kernel export remains contract/test surface while thin CLI is forbidden from importing the kernel barrel.",
-        "until": "2026-09-30"
-      },
-      {
-        "value": "entityRegistryKinds",
-        "ref": "task_01KZRMKJT7YKQ2TNXA8WR1W5DN",
-        "reason": "W3 production flip removed the former CLI barrel consumer; the private kernel export remains contract/test surface while thin CLI is forbidden from importing the kernel barrel.",
-        "until": "2026-09-30"
-      },
-      {
-        "value": "EntityRegistryShape",
-        "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
-        "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
-        "until": "2026-09-30"
-      },
-      {
         "value": "EntityRelationRecordSchema",
         "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
         "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
@@ -457,12 +439,6 @@
         "value": "EntityStorageForm",
         "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
         "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
-        "until": "2026-09-30"
-      },
-      {
-        "value": "entityStorageForms",
-        "ref": "task_01KX7G4G3C4H62Y5SFXZV6RYT2",
-        "reason": "F1 registry substrate declaration contract; scheduled consumers are the F2 Session and F3 Execution/Review slices (dec_mrem0ys8, ADR-0027) — the stale-entry rule will force removal once they land.",
         "until": "2026-09-30"
       },
       {
@@ -586,18 +562,6 @@
         "until": "2026-09-30"
       },
       {
-        "value": "getEntityKind",
-        "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
-        "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
-        "until": "2026-09-30"
-      },
-      {
-        "value": "getEntityRegistration",
-        "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
-        "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
-        "until": "2026-09-30"
-      },
-      {
         "value": "hardFail",
         "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
         "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
@@ -697,12 +661,6 @@
         "value": "isCloseoutReadiness",
         "ref": "task_01KWWCBRSV0V3AWTCM3ZZ1J998",
         "reason": "F8a first enforcement pass preserves the existing kernel public barrel zero-consumption export baseline; future additions must show a consumer or carry new governance evidence.",
-        "until": "2026-09-30"
-      },
-      {
-        "value": "isEntityStorageForm",
-        "ref": "task_01KX7G4G3C4H62Y5SFXZV6RYT2",
-        "reason": "F1 registry substrate declaration contract; scheduled consumers are the F2 Session and F3 Execution/Review slices (dec_mrem0ys8, ADR-0027) — the stale-entry rule will force removal once they land.",
         "until": "2026-09-30"
       },
       {
